added opencv functionality to android framecapture code
added opencv libraries, and framecapture is now successfully displaying
pixel data.

KNOWN ISSUES: drops pixels when sending frame 1 row at a time

+++ sw/opencv_android/FrameCapture/src/glasses/ardueye/framecapture/FrameBufferEvent.java
@@ -7,6 +7,6 @@ public interface FrameBufferEvent {
 	 * 
 	 * @param frame int array containing the frame
 	 */
-	public void onFrameComplete(int[] frame);
+	public void onFrameComplete(byte[] frame);
 	
 }
